[ANDROID-2556] updated annotation classes;
