--- src/main/java/com/tenor/android/core/constant/ContentFormats.java
@@ -1,26 +1,0 @@
-package com.tenor.android.core.constant;
-
-/**
- * The supported content formats
- */
-public class ContentFormats {
-    /**
-     * "image/gif"
-     */
-    public static final String IMAGE_GIF = "image/" + MediaFormats.GIF;
-
-    /**
-     * "image/jpeg"
-     */
-    public static final String IMAGE_JPEG = "image/" + MediaFormats.JPEG;
-
-    /**
-     * "image/png"
-     */
-    public static final String IMAGE_PNG = "image/" + MediaFormats.PNG;
-
-    /**
-     * "image/mp4"
-     */
-    public static final String IMAGE_MP4 = "video/" + MediaFormats.MP4;
-}
