--- src/main/java/com/tenor/android/core/constant/MediaCollectionFormats.java
@@ -1,38 +1,0 @@
-package com.tenor.android.core.constant;
-
-/**
- * The supported media formats
- */
-public class MediaCollectionFormats {
-    /**
-     * Standard GIF format
-     */
-    public static final String GIF = "GIF";
-    /**
-     * Compressed gif.  Good for quick loading and sending
-     */
-    public static final String GIF_TINY = "GIF_TINY";
-    public static final String GIF_MEDIUM = "GIF_MEDIUM";
-    public static final String GIF_NANO = "GIF_NANO";
-
-    /**
-     * Standard mp4 format
-     */
-    public static final String MP4 = "MP4";
-    /**
-     * Compressed mp4 format.  Good for quick loading and sending
-     */
-    public static final String MP4_TINY = "MP4_TINY";
-    public static final String MP4_NANO = "MP4_NANO";
-    /**
-     * Looped mp4 format that contains 3 loops built in
-     */
-    public static final String MP4_LOOPED = "MP4_LOOPED";
-
-    /**
-     * Web player url
-     */
-    public static final String WEBM = "WEBM";
-    public static final String WEBM_TINY = "WEBM_TINY";
-    public static final String WEBM_NANO = "WEBM_NANO";
-}
