--- src/main/java/com/tenor/android/core/constant/MediaFormats.java
@@ -1,13 +1,0 @@
-package com.tenor.android.core.constant;
-
-/**
- * The supported media formats
- */
-public class MediaFormats {
-    public static final String ALL = "all";
-    public static final String GIF = "gif";
-    public static final String MP4 = "mp4";
-    public static final String PNG = "png";
-    public static final String JPG = "jpg";
-    public static final String JPEG = "jpeg";
-}
