+++ src/main/java/com/tenor/android/core/constant/SupportMessenger.java
@@ -4,42 +4,183 @@ import android.support.annotation.StringDef;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
 
-import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.PACKAGE;
-import static java.lang.annotation.ElementType.PARAMETER;
+/**
+ * Contains package name of all supported messengers are sdk can work with.
+ * Additional apps may be added if needed or desired.
+ */
+public class SupportMessenger {
 
-@Retention(RetentionPolicy.CLASS)
-@StringDef({SupportMessengers.NONE,
-        SupportMessengers.FB_MESSENGER,
-        SupportMessengers.WHATSAPP,
-        SupportMessengers.MESSAGES,
-        SupportMessengers.HANGOUTS,
-        SupportMessengers.CHOMP,
-        SupportMessengers.SKYPE,
-        SupportMessengers.EIGHT_SMS,
-        SupportMessengers.TWITTER,
-        SupportMessengers.GO_SMS,
-        SupportMessengers.TANGO,
-        SupportMessengers.KIK,
-        SupportMessengers.WE_CHAT,
-        SupportMessengers.GOOGLE_MESSENGER,
-        SupportMessengers.HIKE,
-        SupportMessengers.HTC_MESSAGES,
-        SupportMessengers.TELEGRAM,
-        SupportMessengers.HOVERCHAT,
-        SupportMessengers.VIBER,
-        SupportMessengers.KAKAO_TALK,
-        SupportMessengers.SLACK,
-        SupportMessengers.VODAFONE,
-        SupportMessengers.FACEBOOK,
-        SupportMessengers.LINE,
-        SupportMessengers.REDDIT,
-        SupportMessengers.GMAIL,
-        SupportMessengers.AOSP_MESSAGES})
-@Target({METHOD, PARAMETER, FIELD, ANNOTATION_TYPE, PACKAGE})
-public @interface SupportMessenger {
+    @Retention(RetentionPolicy.CLASS)
+    @StringDef({NONE,
+            FB_MESSENGER,
+            WHATSAPP,
+            MESSAGES,
+            HANGOUTS,
+            CHOMP,
+            SKYPE,
+            EIGHT_SMS,
+            TWITTER,
+            GO_SMS,
+            TANGO,
+            KIK,
+            WE_CHAT,
+            GOOGLE_MESSENGER,
+            HIKE,
+            HTC_MESSAGES,
+            TELEGRAM,
+            HOVERCHAT,
+            VIBER,
+            KAKAO_TALK,
+            SLACK,
+            VODAFONE,
+            FACEBOOK,
+            LINE,
+            REDDIT,
+            GMAIL,
+            AOSP_MESSAGES})
+    public @interface Value {
+    }
+
+    /**
+     * Empty string, ""
+     */
+    public static final String NONE = StringConstant.EMPTY;
+
+    /**
+     * Facebook Messenger
+     */
+    public static final String FB_MESSENGER = "com.facebook.orca";
+
+
+    /**
+     * WhatsApp Messenger
+     */
+    public static final String WHATSAPP = "com.whatsapp";
+
+    /**
+     * Android Default Messages
+     * Samsung Messages
+     */
+    public static final String MESSAGES = "com.android.mms";
+
+    /**
+     * Hangouts
+     */
+    public static final String HANGOUTS = "com.google.android.talk";
+
+    /**
+     * chomp SMS
+     */
+    public static final String CHOMP = "com.p1.chompsms";
+
+    /**
+     * Skype
+     */
+    public static final String SKYPE = "com.skype.raider";
+
+    /**
+     * Eight sms
+     */
+    public static final String EIGHT_SMS = "com.thinkleft.eightyeightsms.mms";
+
+    /**
+     * Twitter
+     */
+    public static final String TWITTER = "com.twitter.android";
+
+    /**
+     * GO SMS Pro
+     */
+    public static final String GO_SMS = "com.jb.gosms";
+
+    /**
+     * Tango
+     */
+    public static final String TANGO = "com.sgiggle.production";
+
+    /**
+     * Kik
+     */
+    public static final String KIK = "kik.android";
+
+    /**
+     * WeChat
+     */
+    public static final String WE_CHAT = "com.tencent.mm";
+
+    /**
+     * Messenger
+     */
+    public static final String GOOGLE_MESSENGER = "com.google.android.apps.messaging";
+
+    /**
+     * hike messenger
+     */
+    public static final String HIKE = "com.bsb.hike";
+
+    /**
+     * Messages
+     */
+    public static final String HTC_MESSAGES = "com.htc.sense.mms";
+
+    /**
+     * Telegram
+     */
+    public static final String TELEGRAM = "org.telegram.messenger";
+
+    /**
+     * HoverChat
+     */
+    public static final String HOVERCHAT = "com.ninja.sms.promo";
+
+    /**
+     * Viber
+     */
+    public static final String VIBER = "com.viber.voip";
+
+    /**
+     * KakaoTalk
+     */
+    public static final String KAKAO_TALK = "com.kakao.talk";
+
+    /**
+     * Slack
+     */
+    public static final String SLACK = "com.Slack";
+
+    /**
+     * Vodafone Message+
+     */
+    public static final String VODAFONE = "com.vodafone.messaging";
+
+    /**
+     * Facebook
+     */
+    public static final String FACEBOOK = "com.facebook.katana";
+
+    /**
+     * LINE
+     */
+    public static final String LINE = "jp.naver.line.android";
+
+    /**
+     * Reddit
+     */
+    public static final String REDDIT = "com.reddit.frontpage";
+
+    /**
+     * Gmail
+     */
+    public static final String GMAIL = "com.google.android.gm";
+
+    /**
+     * AOSP Messages
+     */
+    public static final String AOSP_MESSAGES = "com.android.messaging";
+
+    /**
+     * Samsung Messaging
+     */
+    public static final String SAMSUNG_MESSAGING = "com.samsung.android.messaging";
 }
