--- src/main/java/com/tenor/android/core/constant/ViewActions.java
@@ -1,8 +1,0 @@
-package com.tenor.android.core.constant;
-
-public class ViewActions {
-
-    public static final String VIEW = "view";
-    public static final String SHARE = "share";
-    public static final String TAP = "tap";
-}
